Move enums to own class
