+++ osu.Framework/Graphics/BlendingParameters.cs
@@ -314,57 +314,4 @@ namespace osu.Framework.Graphics
 
         public override string ToString() => $"BlendingParameter: Factor: {Source}/{Destination}/{SourceAlpha}/{DestinationAlpha} RGBEquation: {RGBEquation} AlphaEquation: {AlphaEquation}";
     }
-
-    public enum BlendingType
-    {
-        Inherit = 0,
-        ConstantAlpha,
-        ConstantColor,
-        DstAlpha,
-        DstColor,
-        One,
-        OneMinusConstantAlpha,
-        OneMinusConstantColor,
-        OneMinusDstAlpha,
-        OneMinusDstColor,
-        OneMinusSrcAlpha,
-        OneMinusSrcColor,
-        SrcAlpha,
-        SrcAlphaSaturate,
-        SrcColor,
-        Zero
-    }
-
-    public enum BlendingEquation
-    {
-        /// <summary>
-        /// Inherits from parent.
-        /// </summary>
-        Inherit = 0,
-
-        /// <summary>
-        /// Adds the source and destination colours.
-        /// </summary>
-        Add,
-
-        /// <summary>
-        /// Chooses the minimum of each component of the source and destination colours.
-        /// </summary>
-        Min,
-
-        /// <summary>
-        /// Chooses the maximum of each component of the source and destination colours.
-        /// </summary>
-        Max,
-
-        /// <summary>
-        /// Subtracts the destination colour from the source colour.
-        /// </summary>
-        Subtract,
-
-        /// <summary>
-        /// Subtracts the source colour from the destination colour.
-        /// </summary>
-        ReverseSubtract,
-    }
 }
